色の計算をコンポーネント内に移動
